pyproject.toml
[project]
name = "invenioscan"
version = "0.1.0"
description = "FastAPI middleware and mobile ingestion tooling for InvenioILS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.115.12",
	"httpx>=0.28.1",
	"pydantic-settings>=2.8.1",
	"pyjwt>=2.10.1",
	"qrcode[pil]>=8.2",
	"uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.3.5",
	"pytest-asyncio>=0.26.0",
	"ruff>=0.11.2",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
